fixed dedupe, and logs
