+++ lib/cache/cacheFactory.js
@@ -8,27 +8,9 @@ var utils = require('../utils');
 var getCacheKey = utils.getCacheKey;
 var getRedisConfig = utils.getRedisConfig;
 
-function hasCacheControl(res, value) {
-    return (res.headers['cache-control'] || '').indexOf(value) !== -1;
-}
-
-function getCacheValidity(args, res) {
-    var options = args[0];
-    if (options.url === 'cache') {
-        return 0;
-    }
-    if (hasCacheControl(res, 'no-cache') || hasCacheControl(res, 'no-store')) {
-        return 0;
-    }
-    if (hasCacheControl(res, 'max-age')) {
-        return parseInt(res.headers['cache-control'].split('=')[1], 10);
-    }
-    return options.hasOwnProperty('cacheTTL') ? (options.cacheTTL/1000) : 60;
-}
-
 var cleanupPeriod = 60 * 60 * 24 * 3; // 3 days (or a long weekend)
 function getCacheTTL(args, res) {
-    return getCacheValidity(args, res) ? cleanupPeriod : 0;
+    return utils.getCacheValidity(args, res) ? cleanupPeriod : 0;
 }
 
 function serialize(res) {
@@ -67,7 +49,7 @@ module.exports = function cacheFactory(config) {
     cache = cacheManager.caching(CMConfig);
     return new Cache(cache, {
       key: getCacheKey(config),
-      maxValidity: getCacheValidity,
+      maxValidity: utils.getCacheValidity,
       maxAge: getCacheTTL,
       serialize: serialize,
       deserialize: deserialize,
