+++ lib/utils.js
@@ -32,6 +32,24 @@ function getCacheKey(config) {
 	}, namespace);
 }
 
+function hasCacheControl(res, value) {
+    return (res.headers['cache-control'] || '').indexOf(value) !== -1;
+}
+
+function getCacheValidity(args, res) {
+    var options = args[0];
+    if (options.url === 'cache') {
+        return 0;
+    }
+    if (hasCacheControl(res, 'no-cache') || hasCacheControl(res, 'no-store')) {
+        return 0;
+    }
+    if (hasCacheControl(res, 'max-age')) {
+        return parseInt(res.headers['cache-control'].split('=')[1], 10);
+    }
+    return options.hasOwnProperty('cacheTTL') ? (options.cacheTTL/1000) : 60;
+}
+
 function filterHeaders(h) {
     var blacklistedHeaders = ['set-cookie'];
 		var filteredHeaders;
@@ -77,5 +95,6 @@ module.exports = {
 	urlToCacheKey: urlToCacheKey,
 	cacheKeytoStatsd: cacheKeytoStatsd,
   filterHeaders: filterHeaders,
-	getRedisConfig: getRedisConfig
+	getRedisConfig: getRedisConfig,
+	getCacheValidity: getCacheValidity
 };
